BSIS-578 Prioritise NOT_DONE blood typing status over PENDING_TESTS
The logic to decide whether or not a test batch can be released is based
on the blood typing status of every donation being NOT_DONE. With the
change to make Titre and AbScr tests mandatory it became possible to go
from NOT_DONE to PENDING_TESTS and back to NOT_DONE which breaks the
behaviour of when a test batch can be released.

This change makes sure that the blood typing status of a donation
remains as NOT_DONE until *all* of the initial serology tests are
completed, instead of changing to PENDING_TESTS when there are pending
repeat tests as well as pending initial tests. This prevents the test
batch from being released until all intial tests are done.