BSIS-639 Add periodic flushes and clears
Periodically flush the hibernate session to the database and then clear
it to prevent slow down. This drastically improves the speed of
importing the donors but has less effect on the donations as a flush is
already being done after saving each donation and there is more logic
being run. Importing the other sheets was reasonably fast already.