BSIS-637 Fix donation batch grouping
The logic to group donation batches was including the time part of the
date. This change groups donation batches by the date part only so that
all donations on the same day are grouped together regardless of time.
The change to the spreadsheet adds time components to the first two
donations in order to test this case.