BSIS-846 Change endpoint to accept a list of blood typing resolutions
Changed the path to "/donations/bloodTypingResolutions" and the method
to POST. Accepting a list of blood typing resolutions as the request
body was not allowing the validator to be run. The solution was to
create a wrapper backing form which has a validator to validate the
nested list of backing forms. The endpoint now accepts an object with a
single field "bloodTypingResolutions" which is the array of blood typing
resoltion backing forms.