BSIS-832: Create ComponentBatchBackingForm
First draft of the ComponentBatchBackingForm. The form only contains
the data that the user will capture (based off the mock-up on the
user story). Other data required for ComponentBatch will need to be
set by the backend based on the DonationBatch selected. (E.g.
status, location, collectionDate, bloodTransportBoxes).

The form can be extended later if editing is necessary.