BSIS-847 Add donationBatch to ComponentBatch entity
This should have been a OneToOne relationship but we couldn't manage to solve an error saying "cannot simultaneously fetch multiple bags". The solution was to use a set of DonationBatch objects instead of just one donationBatch.

This should ideally be a temporary solution until we find a way of implementing the OneOnOne relationship without errors.