BSIS-835: Add query to find unassigned DonationBatches
In order to find the DonationBatches that are available for selection
when creating a ComponentBatch, a query is needed that identifies
DonationBatches that have Donations with Components that have not
been assigned to a ComponentBatch.

It is possible to avoid using the DonationBatch - ComponentBatch
association and rather use the Component - ComponentBatch association
but there was an issue with adding a second join. This should be possible
and if we want to remove the DonationBatch mapping at a later stage
it will be possible to simply modify the query.