BSIS-913 Fix query for finding locations by type
The query was not considering the fact that a location could be more
than one type (e.g. a venue and a processing site) and was setting one
flag to true and the others to false when in fact they should not have
been set. This change uses a separate query to match for each type.