BSIS-914: Fix issue with audit fields for BloodTransportBox
The EntitySaveListener was not triggered on cascaded persists or
merge because not all the methods were implemented. I have implemented
the other onPersist and onMerge methods that are called for all
associated entities when they are saved.

I also added a check not to overwrite the createdBy attribute if
onPersist is called on an entity that has already been saved. It
appears that if persist is done on the original object, a persist
will be done on the associated entities that have been modified.