BSIS-896 Use LocationBackingForm instead of Location in other backing forms
The accessor methods on location are named differently to those on the
view model and backing form and this was causing issues when nested
locations were being used in other backing forms. This change wraps
locations used in other view models in LocationBackingForms so that the
same methods are used for getting and setting the values.