BSIS-948: Created OrderFormComponent entity
Added the OrderFormComponent entity instead of using Hibernate @JoinTable
to join two entities in a many to many relationship. The reason for this
is because the audit table and tracking columns will not be populated
without an entity.

Also added a primary key and composite foreign key reference
in OrderForm_Component table. The composite key is used to ensure
that the same Component is not added twice. (reverted from commit 18f35847a5b806d5d5b0c904344b5582bf3ea152)