BSIS-1052 Fix random failing test: testExpiryStatusWithFutureExpiryDate_shouldReturnDaysUntilExpiry
There is a "days between" calculation between new Date() and a 20 days after, date. The result is sometimes 20 and sometimes 19 depending on the millisecond difference.

This was solved by writing the test with a date where the difference with new Date() is not an exact number of days but 99 hours.