BSIS-1055: Add forms endpoint for Stock Levels report
The inventorystaff user does not have a permission to manage locations,
so they were unable to select a distribution center from the drop down
on the View Stock Levels page. To solve this, I added a forms endpoint
for the stockLevel report that has the same permission as the report.