BSIS-1053 Update Abo/Rh and TTI Reports to use donation venue instead of donor venue in query
Had to update the from part of the query to "FROM Donation d, Donor do" for JPA to correctly create the sql query to return the donation's venue. If this was not done an error was being thrown (Unknown column 'donation2_.venue_id' in 'on clause'). (reverted from commit 7d5918fc327577fb9b69e8539c0b74e626afe50d)