BSIS-1082 Refactor DonationViewModel to not wrap entity
The exception was being thrown because of a circular reference between
component batch and blood transport box. Simply setting the getter for
component batch on blood transport box to json ignored fixed the
original issue, but still threw a lazy initialisation exception.

This refactor removes the donation entity from the donation view model
and instead adds just the fields. This prevents entities from being sent
in the view model which means less data is sent, and there should not be
any circular references.