BSIS-1082 Use donation type backing form in donation backing form
Make sure that the fields returned by the view model match the fields
expected in the backing form so that it's possible to edit a donation.
This also updates the donation form endpoint to return donation type
view models instead of just donation types so that the fields match when
creating a donation.