BSIS-1084: Restructure tests for InventoryRepository StockLevel Restructured the tests for the two InventoryRepository methods to ensure that Components are not being excluded for multiple reasons and so that tests are only testing one condition in the query at a time.
Also added tests to ensure that deleted Components weren't being
returned.

And added extra tests to ensure that the results were being grouped
correctly by blood type.

Note that it's not possible to test ComponentStatus.ISSUED and
ComponentStatus.USED because they should only be used in conjunction
with InventoryStatus.REMOVED which is not expected to be used with
the Stock Levels query. Therefore there are no tests for these
statues.