BSIS-872 Use matchers in tests
When setting up the mocks the methods were expecting an object and doing
an equality comparison on the object to decide if the mocked response
should be returned. The problem with this approach is that the default
equals implementation for the entities just compares the ids, so when
testing that an entity is updated the existing and updated entities were
matching on ids and the other fields were not taken into account. As a
result of this one of the tests was passing incorrectly since the
expectation was being met for the wrong entity.

This change uses matchers to set up the expectations so that the mocked
responses will only be returned if the method is called with an object
with all of the correct fields instead of just a matching id.