BSIS-1289 Add canPrintPackLabel constraint check
Adds two methods to the constraint checker: one to simply do the
constraint check and the second to also do consistency checks. The first
one should only be used for calculating permissions and the second
should be used before actually printing a pack label to make sure that
the data is consistent.