BSIS-1322 Remove getLocationsByType method
There were no usages of this method where a dynamic type was being
passed in so they could all be replaced with the explicit queries. The
LocationType enum is being repurposed for setting the correct booleans
when adding or editing locations so it should no longer be used for
querying.