BSIS-1338 Update createdDate timestamp column definition on TestBatch
After updating testBatch location to the default one, createdDate for all rows was updated to current date for some mysql instances. This happened because for some mysql versions, if all timestamp column default values are not specified for a table, the first tiemstamp column is created with an automatic update for everytime the value of any other column in the row is changed.

This is fixed by updating the definition of createdDate, and set the default to current date.

The liquibase change to do this, is specific to mysql. The syntax to achieve running sql specific for a dbms, was only available for liquibase versions from 3.0, so liquibase was updated to 3.0.0.