BSIS-1347 Upgrade liquibase to version 3.5.1
When installing with a clean db, one of the liquibase (3.0.0) scripts was failing with error: Unknown property autoIncrement for liquibase.datatype.core.SmallIntType smallint(6). Upgrading to liquibase 3.5.1 solved the issue.

Also, for new installations, the field isVenue in table Location, was updated to bit(1) instead of tinyint(1) when the column was renamed (from isDonorPanel). This was causing issues when setting that field. The column type was updated to tinyint(1) to solve this issue.