BSIS-1370 Mark components as unsafe if donor was ineligible
When calculating the status of a component it should be set to unsafe if
the donation was back entered for a donor who was ineligible at the
point of donation. The components are marked as unsafe when the donation
is captured so they should remain unsafe if the status is updated.