BSIS-1382 Include deleted component types in query
When changing the component type code of a component we also need to
check for deleted component types since the code may already be in use.
The constraint will prevent duplicates from being added so the disabled
(deleted) one should be enabled (undeleted) if the code needs to be
used.