BSIS-1437 NT (Unsafe) released donation processed as Available (#591)
* BSIS-1438 Check blood typing match status in updateComponentStatus

If the donation has been released and the blood typing match status is
not one which confirms the blood group then the component should be
marked as unsafe. This matches the behaviour when releasing a donation
with an unconfirmed blood group.

Add additional checks to make sure that the donation to which the
component belongs has been released and has a confirmed blood group
before marking the component as available. This reduces the risk of
components being accidently marked as available.