BSIS-1374 Configure log levels
Leave the root log level as error, but update the log level for the
BSIS package to DEBUG by default. Also update the general config logic
to set the log level for the BSIS package instead of the root log level.
Setting the root level to anything lower than info causes too much noise
and a performance hit due to the volume of log messages, so it makes
more sense to change the BSIS log level.