BSIS-1376: Fix logging in GlobalControllerExceptionHandler Fixed the logging in the global ExceptionHandler. Removed printStackTrace calls which log to System.err and replaced with suitable logging.
Also suppressed the logging of stack traces on ERROR level to reduce
the size of the log file - for errors that occur frequently and
aren't fatal (e.g. NoResultException or MethodArgumentNotValidException).
In those cases the stack trace can be logged on TRACE if a developer
needs to figure out from which line of code the exception is being thrown
from.