BSIS-1484 Prevent hibernate from loading duplicate donation batches
Due to the way that hibernate does joins for collections with a fetch
type of eager it was fetching the same donation batch multiple times
when fetching a test batch. This meant that at the point of releasing a
test batch the same donation batch was being released multiple times. By
changing the list of donation batches to a set we ensure that this can
no longer happen.

Another alternative was to set the fetch mode to SELECT on the
association, but this would result in a separate query being done so we
opted for the set for now.