BSIS-1571 Only check constraints if pack type is being updated
When updating a donation only check the constraints for updating a pack
type when the pack type is being updated. Prior to this the donor
deferral status was being checked when updating any donation field which
was causing an IllegalArgumentException to be thrown even in cases where
it was not applicable.