BSIS-1546 Move endpoint logic to controller service method
Move the code to find a donation by id and convert it to a view model
into a controller service so that it is wrapped in a transaction. The
endpoint is not transactional so a LazyInitializationException was being
thrown when trying to get the components for a donation in the
canEditPackType constraint check.