BSIS-1550 Move addDonation and createInitialComponent logic from repo
Add donation logic was moved to the donation crud service, and createInitialComponent to the component crud service.

By doing this, the donation, donor and components are saved separately and no refresh and flushes are needed.

Also, now we are able to test what is created when calling the createInitialComponent method.