BSIS-1625 Add unit test for DataExportService
Add a unit test to validate the output of the data export service. The
test sets up some data which populates all columns of CSVs and compares
the MD5 sum of the resulting output to the MD5 sum of the expected,
manually validated output. The test fixture uses constant dates to
ensure that the expected MD5 sum does not change between test
executions.