BSIS-1625 Set and unset timezone before and after test
The CSV that the MD5 sum was calculated from was generated in the
Africa/Johannesburg time zone so in order for the dates to be formatted
the same in the test so that the hash doesn't change the same time zone
needs to be used. This change sets the time zone before the test is run
and then resets it to the original value afterwards.