BSIS-1797: Fix Mobile Clinic Export performance
Fixed the performance issue with the Mobile Clinic Export for 'all venues'.
The issue was donor eligibility was determined by retrieving the entire
Donor entity from the database (one by one) and inspecting the donations.

The check (for eligibility on a specific date which is used by the Mobile
Clinic Exports) has been replaced with the query to determine the donor's
next donation date (dueToDonate). If the dueToDonate date is after the
specified date (and the donor doesn't have any active deferrals on that
date) then the donor is eligible.

Note: there are two queries to determine Donor eligibility. One is for
the current date and this method was not changed. These methods were
always implemented independently and the eligibility on the current
date is used when creating a Donation. In the case of a new Donor being
imported with existing Donations (e.g. in the Data Import) the query
would cause a problem, whereas it is possible to retrieve the Donor
from the cache and iterate through the Donations without any performance
impact.