BSIS-1767: Create form endpoint for Adverse Event
Created the form endpoint for Donors Adverse Events Summary Report. This
endpoint needs to return the venues for displaying a drop down containing
venues and a list of the active adverse event types for displaying
the report with variable columns (users can add their own
adverse event types and these should be reflected in the report).