BSIS-1853 Remove unnecessary list of combination component types
The fronted only uses the name and id from the combination view model. The backend was using the list of component types from the combination, instead of fetching the component types from the db.

To improve this, when processing a component, the combination is fetched from the db, and also the list of component types.

The ComponentTypeCombinationSearchViewModel is not needed anymore, and the list of componentTypes was removed from ComponentTypeCombinationViewModel.