BSIS-1880: Merge two changesets into one
Merged two changesets, for creating and initialising ComponentType
tracker fields, into one. (Code review feedback)

Note: the change was done in such a way to minimise the impact on other
developers by adding a precondition to ensure the changeset did not
run a second time and by changing the id to avoid checksum errors.