BSIS-1901 Fix persisting of ComponentTypeCombination
Update to ComponentTypeCombinationFactory createEntity method to
update the Source Component with the new produced Component Type
Combination.

Previously when adding a new ComponentTypeCombination the
ComponentType_ComponentTypeCombination was not udpated with an entry.
This table maps Source component types to produced component type combinations.

The JUnit test cases are also updated with 2 more scenarios added.