BSIS-1909: Fix bug capturing test only donations
Fixed NullPointerException that was happening when adding a donation that
did not produce Components (e.g. Test Only, Dry Pack and Did not bleed)
which meant that it was not possible to create Donations with those
PackTypes.

The NPE was caused by a null initial Component being added to the
Donation which was being traversed during constraint checking.