BSIS-1926: Refactor BloodTestViewModel
1st iteration of the refactor of the BloodTestViewModel to remove the
wrapped entity. This also resulted in refactoring of BloodTestResultViewModel
which was creating an instance of the BloodTestViewModel in a getter.

Removed notes and reagentLotNumber from BloodTestResultViewModel as they
are not used in the frontend and added BloodTestType to BloodTestViewModel
as it is used in the TestResultsController to determine which blood tests
require reEntry and have pending tests (from the BloodTestResultViewModel).