BSIS-1982 Add createEntity method in BloodTestFactory
Add method createEntity to BloodTestFactory in order to transform a BloodTestBackingForm to a BloodTest entity. Add method  ‘createEntity’ to BloodTestFactoryTests to unit test createEntity in BloodTestFactory; also add the corresponding helper classes: BloodTestBackingFormBuilder (to simplify the build of a new instance of  BloodTestBackingForm) and BloodTestMatcher (to assert that 2 instances of BloodTest are similar).