BSIS-330 Create JUnit testcases for calculateOverviewFlags
There was no test coverage for the calculateOverviewFlags(ruleResults)
method, despite the fact that this method has quite a bit of logic.

This submission adds a TestResultControllerTests file to the JUnit test
suite. The endpoint method findTestResultsOverviewForTestBatch is tested because
the calculateOverviewFlags method is private, but is called by the
findTestResultsOverviewForTestBatch endpoint.

The TestResultsController has also been refactored to move common code into
a helper method called getBloodTestingRuleResults.