BSIS-2049: Update to allow PROCESSED & REMOVED
Updated the constraint validator for Component to allow for a Component
Status of PROCESSED and an Inventory Status of REMOVED for the case where
a Component can be re-processed after being labelled (and put in stock).
For example with Fresh Frozen Plasma that can be labelled and processed
further into Cryoprecipitate.