BSIS-2063: Update logic for pending ABO Rh tests
Updated the logic in the BloodTestingRulesEngine so that ABO and Rh tests
are cleared by default and not cleared when the donor is a first time
donor and both ABO and Rh and have been captured.

This fixes an issue where pending ABO/Rh tests were triggered for
a repeat donor, if only ABO or Rh outcomes were entered.