BSIS-2178: Remove liquibase from mvn install
Removed liquibase from the mvn install lifecycle but added it to travis
so it will still run on the build server after the install has been
successful.

This is good because it is useful to run liquibase on a clean database
to ensure that it will always work for new installations.