BSIS-2211: Release Donation when PackType changed
When a PackType change results in a new Initial Component being created
(e.g. changing PackType from test only to one that counts as a donation
such as Quad) the Donation must be released if the TestBatch is either
released or closed.

This ensures that the Component has the correct status, and in the event
that the Component is unsafe, the correct ComponentStatusChange history
will be generated.