BSIS-2207 Add version info to MANIFEST.MF
Update POM file to include POM version and git describe
value in the MANIFEST.MF file. This is to be used to
display the backend version information on the frontend in
future.

This commit updates the POM to make use of the
gitdescribe-maven-plugin to make the git-describe value
available as a maven env variable i.e. {describe} to be
used by the maven-war-plugin. It also updates the config
of the maven-war-plugin to include Default Implementation
Entries in the MANIFEST.MF file.