BSIS-2215 Remove database and liquibase prop files
Removed the database.properties and liquibase.properties files because
the properties are also in bsis.properties and it's confusing and
unneccessary to have multiple properties files.

Created a test bsis.properties file (and saved in the test resources
folder so it is available to the tests) so that we didn't need to have
test database properties in a file that is included in the war.

Also moved the Spring application context file into the test resources
file so that it doesn't need to be included in the war. This could also
solve the problem with duplicate Application Contexts in the classpath
when running Spring unit tests.