BSIS-2216: Remove duplicate db properties
Removed the duplicate database properties in bsis.properties because
it is possible to use the same properties for Liquibase and Hibernate/JPA.

This will make the setup for implementers easier and less error prone
because there is no chance to only update one database setting and
not the other.

Note: that it is not possible to configure Liquibase database property
keys, so unfortunately we had to drop the "database" group for the
driver, url, username and password.