BSIS-2272 BSIS-2090: Remove uninterpretable fields
Removed all the uninterpretable fields from the BloodTestingRuleResultSet
and BloodTestingRuleResult as they are not used. They were previously
used when BSIS did analysis on the test results entered, but we now
only record the final outcomes.

After adding TITRE to DonationField, the Rules Engine was outputting
a warning that Titre was an unknown DonationField, so removing the
unnecessary code was better than suppressing the log.