BSIS-2165 Update pack label ZPL to include flag chars
Previously only the DIN was included in the pack label and
the interpretation line was included as part of the bar code. In
this commit the ^BC interpretation line is turned off and manually
placed instead to allow for the addition of the rotated flag characters
and the boxed check character.

The flag characters are added to the actual bar code thus making the
bar code the DIN + flag characters. The check characters is
not part of the bar code.

The module width of the bar code has been set to 2 and the packing
mode of the ^BC set to Auto. The latter is done so that most DINs
up to length 20 will fit into the top left quadrant.