BSIS-2165 Remove unnecessary type conversions
Previously the positions calculated for the interpretation
line was being calculated, converted to a String and then
converted to an integer, and then finally converted to
a String again.

This seems somewhat absurd, and so now the values are stored
as integers and then converted to Strings when building up
the DIN related ZPL text.