BSIS-2331: Add gravity defaults for ComponentType
Added gravity defaults to the liquibase changesets used to insert default
data for ComponentType. This will only run on empty databases and will
be 'marked as run' on existing databases due to the preconditions.

The previous changesets to insert default ComponentTypes (and related
entities) were moved to the new update file so that they are run after the
gravity column is created.