BSIS-2370: Update createdOn with bleedStartTime
Updated Component.createdOn when the Donation bleedStartTime is updated
(which can be done before test results are added using edit Donation or
during pre-processing of a Component). This is necessary because
the initial component's createdOn date/time is the donation date plus
the starting time of the bleed - and will be used to calculate time since
donation (to determine if a Component is unsafe).

Added javadocs for DonationCRUDService.updateDonation because it's not
possible to properly use this method passing in a managed entity because
the 'existingDonation' retrieved via a query in the updateDonation method
will be the same object as the parameter passed in. This means that the
comparison checks between existing and updated Donation will not work
(and the constraints will not be called).