BSIS-2370: Fix issues with donationDateTime field
Fixed the various issues with adding donationDateTime to the Component
ManagementViewModel. Added JsonSerialize so that the date is outputted
correctly in the JSON, and JsonIgnore setters in the backing forms to
avoid errors where an object created from a view model is posted back
to the server (with unknown fields).