BSIS-2441: Fix issue with updating DonationType
Fixed the issue with updating DonationType where the DonationType generated
from the JSON sent by the frontend was being used to overwrite the entity
in the database. Because the JSON didn't contain the tracking fields, this
caused an exception with the not-null constraint for createdBy.

To solve this issue, a method was added to the DonationTypeCRUDService that
handles updating the DonationType by first retrieving the existing
DonationType from the database and updating the name and the isDeleted
flag.