BSIS-2517 Create method findSafeComponents in LabellingService
This method differentiates when DIN is present and the bloodGroups are converted from a list of String to a list of BloodGroup.

When DIN is not null, is calling findComponentsByDonationIdentificationNumber at the moment, this should be changed to findSafeComponentByDINAndCode, when it's ready. Also the test for that case should be updated.