BSIS-2569 Avoid clearing donation notes when clearing counselling status
The business logic to clear the necessary fields when clearing the couselling status, sits now in the validator, as it validates that if flagged for couselling is true, then the date and status must be null.

There was a method in the crud service that was clearing the fields manually before, and it was setting the donation notes to null, as per previous logic.

The method and tests to manually clear the couselling status have been removed.