BSIS-2597 Use DonationViewModel instead of DonationSummaryViewModel for testBatches
When returning all donations for a test batch, the DonationSummaryViewModel was being used, but it was just returning a DonationViewModel.

Removed the creation of those view models from the donation summary factory to the test batch factory, as getting the donations for a test batch was only used for the test batches section.