BSIS-2598 Fetch Full Location in PDC BackingFormValidator
Previously the PostDonationCounsellingBackingFormValidator was calling
getIsReferralSite on the referall Site, but this was not the full object,
but rather a Management object so only the site ID was set, and not the
site booleans.

This commit updates the code to take the ID, and fetch the actual location
using the LocationRepository.