BSIS-2687 Set created date for reaction types defaults
Some versions of Ubuntu setup mysql by default in a way that TIMESTAMP is always populated with current timestamp. That's why even though createdDate is not nullable in TransfusionReactionType, and it's not set in the default inserts, there's no error when testing (using Ubuntu). But in case this default is not set, createdDate should be set in the inserts.