BSIS-2843 Allow for NoResult when finding OrderForm
Previously, if the component was not be added to any order form
then an error would be returned to the frontend.

This replaces the use of getSingleResult() with getResultList()
and returns the first element. If there are no elements the
findByComponent returns null. If there are more than one Order Form
returned, an exception is thrown. However this should not occur.

There was also an issue with the InventoryFullViewModelMatcher where
a bracket missing around an OR condition causing the last few checks
to be ignored.