BSIS-2848: Fix issue with display of unit gap
Fixed issue with display of the order form units ordered gap. The problem
was with the differences between the InventoryViewModel and the ComponentViewModel.
We changed the frontend to use an InventoryFullViewModel when adding new
Components to the Order Form, but when retrieving the OrderForm, we
returned ComponentViewModels.

This change ensures that we are consistant with the usage of InventoryViewModels
in the OrderForm Inventory section of the app.

It also changed the use of LocationFullViewModel in the InventoryViewModel which
was unnecessary as we only need the Location id and name.