BSIS-2893 Update validation of dateTransfused to only check date
Previously the time part of the transfusedDate was also being
compared with the component's createdOn date. This led to
issues when recording transfusions that was on the same date
that the component was created On.

The check only compares the date portion now by calling a
new DateGeneratorService to convert the dates to a Joda LocalDate
equivalent.