BSIS-2902 Fix conversion from byte(16) to java.util.UUID
Previously UUID.nameUUIDFromBytes(..) was used to convert from the byte
array to the UUID object. This was incorrect, because this returns
a version 3 UUID.

A new method is added to convert from a byte[16] array. This converts
the byte array to two longs, and then passes this to a UUID constructor.

This solution was taken from this stackoverflow answer:
http://stackoverflow.com/a/29836273