BSIS-2911 Convert Division id from Bigint to binary(16) - UUID
Previously auto increment ID was used for the Division table. This
commit changes the type to binary(16) so that a UUID can be used
as the primary key.

This includes all updates to the code to reference the UUID type instead
of the Long type.

Note after this change two test cases are failing. This still needs to be
investigated. They are in org.jembi.bsis.model.DivisionRoundTripTests

* testPersistValidDivisionWithNoName_shouldThrow
* testPersistValidDivisionWithBlankName_shouldThrow