BSIS-2935 Add virtual columns for uuid blobs
This commit adds virtual (generated) columns for all
blob columns. These columns are generated and not stored
and are added to simplify debugging. Columns are added
for all uuid fields relating to the Division table. These
include the Location table foriegn key references and the AUD
tables.

The virtual columns have the same name as the uuid id columns,
but have '_text' appended to them. The value calculated
is the same as what is displayed in the url on the front-end.

This commit requires that the host machine mysql version is upgraded
to mysql 5.7.17