BSIS-2887: Fix Liquibase changesets for timestamps
Fixed the Liquibase changesets for timestamps because the new version of
MySQL doesn't support zero dates. An error is thrown when adding a
'not null' constraint to a timestamp column if the column does not have
a default set. Older version of MySQL automatically used a zero date.

See the ticket for more information.