BSIS-2911 Force Test Persisters to flush
When Entities with auto increment ids are persisted, they
are inserted to the DB early by the hibernate core code,  resulting
in contraint validation exceptions being thrown immediately.

When the entities are updated to use a UUID id, they are not
inserted early, but instead an Insert Action is added to the
Hibernate ActionQueue.

See AbstractEntityInsertAction.isEarlyInsert() implementation.

As a result, the Division Round Trip Tests fail because the test
case ends before the ActionQueue events are processed. This commit
adds a entityManager.flush() to force the execution of events on the
queue before the test case ends. Changing this highlighted some minor
issues in the DivisionRoundTripTests and
ComponentTypeCombinationRepositoryTests files.