BSIS-2217: Re-organise JSP and HTML files
Re-organised the JSP and HTML files so that they are in the root of the
war folder and are easy to find.

Also eliminated extra files by merging index.html and api.html - both
weren't necessary.

Replaced login.jsp with login.html and removed the LoginController
which is no longer required.

Renamed errorPage.jsp to error.html since it only contained HTML.

NOTE: login.html still needs to be styled