BSIS-2957 Fix liquibase UUID update
Added a precondition to check that the primary key is not BINARY
already, before running the changeset. If the primary key is binary
it means the table has been upgraded already.

Also update the changeset to use the new approach of altering the
type of the existing primary key as opposed to creating a new
separate primary key while dropping the old one.