BSIS-2957 Fix Liquibase Changeset for Division UUID updates.
Remove drop and add of primary key. This is not needed.

Add ';' at the end of all sql statements.

Remove any references to bsis schema in the liquibase changeset,
and therefore also remove the prerequisite which queried the
type of the Division id under the bsis schema. This is because
the bsis schema is not relevant for databases that are not
named 'bsis'.