BSIS-2961 Use time as part of UUID used for HSQL
The UUID we generate via mysql is sequential based on the
time part. The UUID we used for HSQL (used in JUnits) was totally
random. This update uses the System current time as the MSB to make
it time sequential, and also negates the leading bit so that the UUID
long value is > 0.