BSIS-2987: Fix failing test for DonationBatch
Fixed failing test for DonationBatchRepository which was failing when
the DonationBatches returned in the List were in a different order to
what was expected.

This should be investigated further, because BSISUUIGenerator should
be generating sequentual uuids for testing, so the old natural order
should be preserved.

For this test, it is OK to use the 'hasItem' matcher because there
is no explict ordering in the query.