BSIS-3036 Pass correct BloodTest to Update Method
Previously the incorrect BloodTest instance was passed
into the update method. The BloodTestCRUDService.updateBloodTest(...)
method calls the bloodTestRepository.update(...) method.

The existing BloodTest, updated with the new values, is now
passed into this method, as opposed to the bloodTest passed into
the BloodTestCRUDService.updateBloodTest(...) method.

This commit also contains a refactor of the BloodTestBuilder
to allow for the building with tracking fields. It's created
in a generic way so that other builders can also leverage it when
needed.