BSIS-3054 Fix non-deterministic OrderFormRepository Test
testFindOrderFormsByPeriodAndLocation_shouldReturnRightOrders was failing
randomly because of the new builder RandomTestDate implementation. One
of the Order forms did not have a date set before persisting.

The RandomTestDate caused the test case to fail when the random
date falls out of range. This update sets the date excplicitly in the
test